feat: :sparkles: implementacion de typescript (@ts-check)

+++ src/js/admin.liga.js
@@ -1,11 +1,22 @@
-import { emptyArray } from './decorators/emptyArray.js'
+// @ts-check
+
 import { dataStore } from './classes/Store.js'
 import { Partido } from './classes/Partido.js'
 import { Jornada } from './classes/Jornada.js'
 import { Liga } from './classes/Liga.js'
+import { Equipo } from './classes/Equipo'
+
+/**
+ * @typedef storedData
+ * @property {Equipo[]=} equipos 
+ * @property {import("./classes/Usuario").Usuario[]=} usuarios 
+ * @property {import("./classes/Noticia").Noticia[]=} noticias 
+ * @property {Liga[]=} ligas 
+ */
 
 document.addEventListener('DOMContentLoaded', onDOMContentLoaded)
 
+/** @type {Equipo[]} */
 const equiposLiga = []
 
 // ------- EVENTS ------- //
@@ -15,9 +26,9 @@ function onDOMContentLoaded() {
     const crearLigaBtn = document.getElementById('crear-liga-btn')
     const clearFormBtn = document.getElementById('clear-form-btn')
 
-    addEquipoBtn.addEventListener('click', addEquipos)
-    crearLigaBtn.addEventListener('click', crearLiga)
-    clearFormBtn.addEventListener('click', clearLigaForm)
+    addEquipoBtn?.addEventListener('click', addEquipos)
+    crearLigaBtn?.addEventListener('click', crearLiga)
+    clearFormBtn?.addEventListener('click', clearLigaForm)
 
     loadEquiposInSelect()
     getLigas()
@@ -29,15 +40,24 @@ function onDOMContentLoaded() {
 
 // ------- METHODS ------- //
 
+/**
+ * Añade equipos a la liga
+ */
 function addEquipos() {
     const select = document.getElementById('sel-equipo')
-    const selectId = select.value
+    const selectId = select?.getAttribute('value') || ''
 
-    const equipo =  dataStore.get().equipos.find(eq => eq.id === selectId)
-    equiposLiga.push(equipo)
-    drawEquipoRow(equipo)
+    const equipo =  dataStore.get().equipos?.find(/**@param {Equipo} eq*/eq => eq.id === selectId)
+    if (equipo) {
+        equiposLiga.push(equipo)
+        drawEquipoRow(equipo)
+    }
 }
 
+/**
+ * Muestra el equipo en la tabla
+ * @param {Equipo} equipo
+ */
 function drawEquipoRow(equipo) {
     const tbody = document.getElementById('tbody-equipos')
     const row = document.createElement('tr')
@@ -50,7 +70,7 @@ function drawEquipoRow(equipo) {
     const delBtn = document.createElement('button')
 
     row.id = `row_${equipo.id}`
-    tbody.appendChild(row)
+    tbody?.appendChild(row)
     cellId.innerText = equipo.id
     row.appendChild(cellId)
     cellNombre.innerText = equipo.nombre
@@ -63,24 +83,34 @@ function drawEquipoRow(equipo) {
     row.appendChild(cellEstadio)
     row.appendChild(cellEdit)
     delBtn.innerText = '🗑'
-    delBtn.addEventListener('click', borrarEquipo.bind(this, equipo.id))
+    delBtn.addEventListener('click', borrarEquipo.bind(delBtn, equipo.id))
     cellEdit.appendChild(delBtn)
 }
 
+/**
+ * Elimina equipo de la liga
+ * @param {string} id 
+ */
 function borrarEquipo(id) {
-    const index = equiposLiga.find(equipo => equipo.id === id)
+    const index = equiposLiga.findIndex(equipo => equipo.id === id)
 
-    document.getElementById(`row_${id}`).remove()
-    equiposLiga.splice(index, 1)
+    document.getElementById(`row_${id}`)?.remove()
+    if (index != -1) {
+        equiposLiga.splice(index, 1)
+    }
 }
-
+    
+/**
+ * Creamos la liga, con sus jornadas y partidos
+ */
 function crearLiga() {
-    const nombreLiga = document.getElementById('nombre').value
-    const yearLiga = document.getElementById('year').value
+    const nombreLiga = document.getElementById('nombre')?.getAttribute('value') || ''
+    const yearLiga = document.getElementById('year')?.getAttribute('value') || ''
     const equipos = [...equiposLiga]
 
     if (equipos.length % 2 != 0) {
-        equipos.push({id: '0000', nombre: 'DESCANSO'})
+        const descanso = new Equipo('DESCANSO')
+        equipos.push(descanso)
     }
     const calendario = new Array(equipos.length-1).fill(null).map(() => new Array(equipos.length-1))
     for (let i = 0; i < equipos.length; i++) {
@@ -107,9 +137,9 @@ function crearLiga() {
                 vuelta.push(new Partido(calendario[i][j], calendario[i][equipos.length - j - 1]))
             }
         }
-        const jornadaClass = new Jornada(i + 1, jornada)
+        const jornadaClass = new Jornada(i + 1, new Date(), jornada)
         jornadas.push(jornadaClass)
-        const vueltaClass = new Jornada(i + equipos.length, vuelta)
+        const vueltaClass = new Jornada(i + equipos.length, new Date(), vuelta)
         jornadasVuelta.push(vueltaClass)
         esLocal = !esLocal
     }
@@ -118,13 +148,17 @@ function crearLiga() {
     const ligaClass = new Liga(nombreLiga, yearLiga, equipos, liga)
     console.log(ligaClass)
 
-    dataStore.get().ligas.push(ligaClass)
+    dataStore.get().ligas?.push(ligaClass)
     localStorage.setItem('storedData', JSON.stringify(dataStore.get()))
    
     drawLigaRow(ligaClass)
     clearLigaForm()
 }
 
+/**
+ * Muestra la liga en la tabla
+ * @param {Liga} liga
+ */
 function drawLigaRow(liga) {
     const tbody = document.getElementById('tbody-ligas')
     const row = document.createElement('tr')
@@ -136,7 +170,7 @@ function drawLigaRow(liga) {
     const delBtn = document.createElement('button')
     
     row.id = `liga_${liga.id}`
-    tbody.appendChild(row)
+    tbody?.appendChild(row)
     cellId.innerText = liga.id
     row.appendChild(cellId)
     cellNombre.innerText = liga.nombre
@@ -145,50 +179,67 @@ function drawLigaRow(liga) {
     row.appendChild(cellYear)
     row.appendChild(cellBtn)
     editBtn.innerText = '✎'
-    editBtn.addEventListener('click', editarLiga.bind(this, liga.id))
+    editBtn.addEventListener('click', editarLiga.bind(editBtn, liga.id))
     cellBtn.appendChild(editBtn)
     delBtn.innerText = '🗑'
-    delBtn.addEventListener('click', borrarLiga.bind(this, liga.id))
+    delBtn.addEventListener('click', borrarLiga.bind(delBtn, liga.id))
     cellBtn.appendChild(delBtn)
 }
 
+/**
+ * muestra la informacion de la liga
+ * @param {string} id 
+ */
 function editarLiga(id) {
-    const liga = dataStore.get().ligas.find(lg => lg.id === id)
+    const liga = dataStore.get().ligas?.find(/** @param {Liga} lg*/lg => lg.id === id)
     const boxJornadas = document.getElementById('box-jornadas')
 
-    document.getElementById('id-liga').value = liga.id
-    document.getElementById('nombre').value = liga.nombre
-    document.getElementById('year').value = liga.year
+    if (liga){
+        document.getElementById('id-liga')?.setAttribute('value', liga.id)
+        document.getElementById('nombre')?.setAttribute('value', liga.nombre)
+        document.getElementById('year')?.setAttribute('value', liga.year)
 
-    liga.equipos.forEach(equipo => drawEquipoRow(equipo))
-    boxJornadas.innerHTML = ''
-    liga.jornadas.forEach(jornada => drawJornadaBox(jornada))
+        liga.equipos.forEach(/** @param {Equipo} equipo */equipo => drawEquipoRow(equipo))
+        if (boxJornadas) boxJornadas.innerHTML = ''
+        liga.jornadas.forEach(/** @param {Jornada} jornada*/jornada => drawJornadaBox(jornada))
+    }
+    
 }
 
+/**
+ * Elimina la liga
+ * @param {string} id 
+ */
 function borrarLiga(id) {
     const ligas =  dataStore.get().ligas
-    const index = ligas.findIndex(liga => liga.id === id)
+    if (ligas) {
+        const index = ligas.findIndex(/** @param {Liga} liga */liga => liga.id === id)
 
-    if (window.confirm(`¿Desea borrar la liga ${ligas[index].nombre}?`)) {
-        ligas.splice(index, 1)
+        if (window.confirm(`¿Desea borrar la liga ${ligas[index].nombre}?`)) {
+            ligas.splice(index, 1)
 
-        document.getElementById(`liga_${id}`).remove()
-        dataStore.get().ligas = ligas
-        localStorage.setItem('storedData', JSON.stringify(dataStore.get()))
+            document.getElementById(`liga_${id}`)?.remove()
+            dataStore.get().ligas = ligas
+            localStorage.setItem('storedData', JSON.stringify(dataStore.get()))
+        }
     }
 }
 
+/**
+ * muestra la jornada
+ * @param {Jornada} jornada 
+ */
 function drawJornadaBox(jornada) {
     console.log(jornada)
     const boxJornadas = document.getElementById('box-jornadas')
     const div = document.createElement('div')
     const title = document.createElement('h3')
 
-    boxJornadas.appendChild(div)
+    boxJornadas?.appendChild(div)
     title.innerText = `Jornada nº: ${jornada.numero}`
     div.appendChild(title)
 
-    jornada.partidos.forEach(partido => {
+    jornada.partidos.forEach(/** @param {Partido} partido */partido => {
         const texto = document.createElement('p')
         texto.innerText = `${partido.local.nombre} vs ${partido.visitante.nombre}`
         div.appendChild(texto)
@@ -196,38 +247,56 @@ function drawJornadaBox(jornada) {
 }
 
 
-
+/**
+ * Carga los equipos en el selector del formulario
+ */
 function loadEquiposInSelect() {
-    const storedData = JSON.parse(localStorage.getItem('storedData')) || {}
+    const jsonStoredData = localStorage.getItem('storedData')
+    /** @type {storedData} */
+    let storedData = {}
+    if (jsonStoredData) {
+        storedData = JSON.parse(jsonStoredData)
+    }
+    //const storedData = JSON.parse(localStorage.getItem('storedData')) || {}
     const select = document.getElementById('sel-equipo')
-    select.innerHTML = `<option value="0">Seleccione un equipo</option>`
+    if (select) select.innerHTML = `<option value="0">Seleccione un equipo</option>`
     if (storedData.hasOwnProperty('equipos')) {
         const equipos = storedData['equipos']
         dataStore.get().equipos = equipos
-        equipos.forEach(equipo => {
-            select.innerHTML += `
+        if (equipos) equipos.forEach(/** @param {Equipo} equipo */equipo => {
+            if (select) select.innerHTML += `
                 <option value="${equipo.id}">${equipo.nombre}</option>
             `
         })
     }
 }
 
+/**
+ * Obtiene las ligas existentes
+ */
 function getLigas() {
-    const storedData = JSON.parse(localStorage.getItem('storedData')) || {}
-    document.getElementById('tbody-ligas').innerHTML = ''
+    const jsonStoredData = localStorage.getItem('storedData')
+    /** @type {storedData} */
+    let storedData = {}
+    if (jsonStoredData) {
+        storedData = JSON.parse(jsonStoredData)
+    }
+    //const storedData = JSON.parse(localStorage.getItem('storedData')) || {}
+    const tbody = document.getElementById('tbody-ligas')
+    if (tbody) tbody.innerHTML = ''
     if (storedData.hasOwnProperty('ligas')) {
         const ligas = storedData.ligas
         dataStore.get().ligas = ligas
-        ligas.forEach(liga => drawLigaRow(liga))
+        if (ligas) ligas.forEach(liga => drawLigaRow(liga))
     }
 }
 
 function clearLigaForm() {
     const select = document.getElementById('sel-equipo')
 
-    select.innerHTML = `<option value="0">Seleccione un equipo</option>`
-    document.getElementById('nombre').value = ''
-    document.getElementById('year').value = ''
-    document.getElementById('tbody-equipos').innerHTML = ''
-    document.getElementById('box-jornadas').innerHTML = ''
+    if (select) select.innerHTML = `<option value="0">Seleccione un equipo</option>`
+    document.getElementById('nombre')?.setAttribute('value', '')
+    document.getElementById('year')?.setAttribute('value', '')
+    document.getElementById('tbody-equipos')?.setAttribute('value', '')
+    document.getElementById('box-jornadas')?.setAttribute('value', '')
 }
